Add files via upload
This is a command-line blog-help tool confirmed to function as expected in Ubuntu. It converts text files to html and searches files. It could thus be integrated into a bloggers workflow, and some functions could be utilized by a site to search entries.

Given no arguments, it checks for an input text directory and output html directory. If there are none or if the text directory is empty, it creates the directories and prompts the user to enter plain text into the input directory. If there are text files in the appropriate directory, it attempts to convert the text files to html in the output directory. The html could be improved, of course.

Given an argument, the tool searches the text directory for all files which contain the search term. Names of all files that include the term are displayed. If there is no directory or files, nothing is displayed. Quotation marks can be used to search for a string with spaces. This allows author, date, content, and title to be searched for.